+++ src/interfaces/index.ts
@@ -1,0 +1,3 @@
+export interface ParamsInterface {
+  [propName: string]: any;
+}
